# docs/swagger.yaml
basePath: /
definitions:
...
  domain.CreateServiceRequestDTO:
    properties:
      actuation_days:
        items:
          type: string
        minItems: 1
        type: array
      category_id:
        type: integer
      description:
        type: string
      end_time:
        type: string
      name:
        type: string
      offer_since:
        type: string
      photos:
        items:
          type: string
        type: array
      start_time:
        type: string
      value:
        type: string
    required:
    - actuation_days
    - category_id
    - description
    - end_time
    - name
    - offer_since
    - start_time
    - value
    type: object
...
  domain.ServiceCategoryDTO:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  domain.ServiceResponseDTO:
    properties:
      actuation_days:
        items:
          type: string
        type: array
      category:
        $ref: '#/definitions/domain.ServiceCategoryDTO'
      description:
        type: string
      end_time:
        type: string
      id:
        type: string
      name:
        type: string
      offer_since:
        type: string
      photos:
        items:
          type: string
        type: array
      start_time:
        type: string
      value:
        type: number
    type: object
...
  domain.UpdateServiceRequestDTO:
    properties:
      actuation_days:
        items:
          type: string
        type: array
      category_id:
        type: integer
      description:
        type: string
      end_time:
        type: string
      name:
        type: string
      offer_since:
        type: string
      photos:
        items:
          type: string
        type: array
      start_time:
        type: string
      value:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API for the Helptrix helper marketplace platform.
  title: Helptrix API
  version: "1.0"
paths:
...
  /service:
    get:
      description: Lists all services for the authenticated helper user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.ServiceResponseDTO'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List services
      tags:
      - service
    post:
      consumes:
      - application/json
      description: Creates a new service offering for the authenticated helper user
      parameters:
      - description: Create service request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.CreateServiceRequestDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.ServiceResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a service
      tags:
      - service
  /service/{id}:
    delete:
      description: Soft-deletes a service for the authenticated helper user
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a service
      tags:
      - service
    get:
      description: Returns a specific service by its ID for the authenticated helper
        user
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ServiceResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get service by ID
      tags:
      - service
    put:
      consumes:
      - application/json
      description: Updates fields of an existing service for the authenticated helper
        user
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      - description: Update service request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateServiceRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ServiceResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a service
      tags:
      - service
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"